pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.7",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
    "ruff>=0.11.2",
]
